Externalize copy functionality
Gradle would like to deprecate the copy methods on Configuration in Gradle 8.2 as there are many problems with them.  Instead of calling copy, this re-implements the existing copy functionality via public API (and, temporarily reflection).  When Gradle 8.2 is released, the reflection can be removed in favor of new public API.

This retains the current behavior of calling copyRecursive when Gradle < 8.2 is detected.